transfer ingest methods
* transfer ingest method to own class
* adjusted in main method
